backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # For PostgreSQL async support
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
opencv-python==4.8.1.78

# AI/ML for chatbot and analysis
google-generativeai==0.3.2
requests==2.31.0

# Background jobs and scheduling
apscheduler==3.10.4

# File handling and utilities
aiofiles==23.2.0
python-dotenv==1.0.0

# HTTP client for external APIs
httpx==0.25.2

# Data processing
pandas==2.1.3
numpy==1.25.2

# Fuzzy string matching for drug names
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS support
python-cors==0.1.0

# JSON handling
orjson==3.9.10

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
